requirements.txt
streamlit
firebase-admin
python-dotenv
cohere
pandas
